This commit was generated by cvs2svn to compensate for changes in r16091, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@16092 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/c64/plus60k.c
@@ -177,7 +177,8 @@ static int set_plus60k_base(resource_value_t v, void *param)
       case 0xd100:
         break;
       default:
-        log_message(plus60k_log, "Unknown +60K base address $%X.", (long)v);
+        log_message(plus60k_log, "Unknown +60K base address $%X.",
+                    (unsigned int)v);
         return -1;
     }
